+++ src/main/java/de/maxhenkel/camerautils/gui/ThirdPersonCameraScreen.java
@@ -5,11 +5,11 @@ import de.maxhenkel.camerautils.Utils;
 import de.maxhenkel.configbuilder.entry.ConfigEntry;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 
 public class ThirdPersonCameraScreen extends SettingsScreenBase {
 
-    private static final ResourceLocation TEXTURE = ResourceLocation.fromNamespaceAndPath(CameraUtils.MODID, "textures/gui/generic_7.png");
+    private static final Identifier TEXTURE = Identifier.fromNamespaceAndPath(CameraUtils.MODID, "textures/gui/generic_7.png");
 
     private ConfigEntry<Double> offsetX;
     private ConfigEntry<Double> offsetY;